lsm/txval: incident #49 slen ABI fix + first-contact standardness caps
Two more members of the relay reject-flood family (LOG.md 2026-08-26):

- incident #49: utxo_lsm_get's run-hit paths stored slen as 4 BYTES while
  the daemon's consensus-adjacent callers declare unsigned long* -- the
  upper half of the caller's variable kept stale stack garbage, and mempool
  prevout resolution nondeterministically failed "prevout script too large"
  (71/29 flip over 100 identical calls; 0x98000000016 for a 22-byte
  P2WPKH). Every LSM differential test declared the same narrow type as
  the asm store, proving mutual consistency instead of the contract; the
  apply path was protected only by resolving via prefetch/memtable, where
  utxo_get already stored 8 bytes. The contract is now explicitly 64-bit:
  one zero-extended 8-byte store at .lg_found, all eleven narrow callers
  widened, test_utxo_lsm poisons the high half and asserts it clears.
  daemon/utxo_probe_one (new) is the tool that made it deterministic;
  offline proof: 150/150 fresh mempool prevouts, zero anomalies.

- first-contact caps: txval_modern bounded txs at 16 inputs / 16 outputs /
  1 KB stripped size -- real batching txs mass-rejected as "malformed tx".
  Now MV_MAX_IN 2048 / MV_MAX_OUT 4096 / 400 KB, sized from Core's 100 kvB
  standardness bound; mv_tx_t static with per-input zeroing at parse time.
  test_mempool_accept_modern pins a 20-in/60-out tx passing the parse
  stage.

Full make -k test green (181 suites).

+++ asm/tests/test_utxo_live_count.c
@@ -43,7 +43,7 @@ extern long utxo_lsm_del(void* lst, void* u, const unsigned char txid[32], unsig
 extern long utxo_lsm_get(void* lst, void* u, const unsigned char txid[32], unsigned index,
                           unsigned long long* value, unsigned long* height,
                           unsigned long* is_coinbase,
-                          const unsigned char** script, unsigned* slen);
+                          const unsigned char** script, unsigned long* slen);
 extern long utxo_lsm_count(void* lst);
 extern long utxo_lsm_reload(void* lst, void* u);
 extern long utxo_lsm_compact(void* lst);
